% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_shape.R
\name{clean_shape}
\alias{clean_shape}
\title{Clean sf data objects}
\usage{
clean_shape(.data, .crs = wgs84)
}
\arguments{
\item{.data}{spatial sf data object}

\item{.crs}{st_crs formatted projection; defaults to wgs84}
}
\value{

}
\description{
This function uses janitor::clean_names() to standardize headers and sets crs to wgs84 by default.
Data must already by in sf format.
}
\examples{
tx_counties <- clean_shape(tx_counties)
}
