% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ja_base_map.R
\name{ja_base_map}
\alias{ja_base_map}
\title{Create JA base map layer in leaflet}
\usage{
ja_base_map(
  .lon = -95.3103,
  .lat = 29.7752,
  .zoom_level = 9,
  .minZoom = 0,
  .maxZoom = 18,
  .dragging = F,
  .tap = F
)
}
\arguments{
\item{.lon}{Longitude for setting center of map}

\item{.lat}{Latitude for setting center of map}

\item{.zoom_level}{Zoom level for setting map}

\item{.minZoom}{Set minimum Zoom level for map}

\item{.maxZoom}{Set max zoom level for map}

\item{.dragging}{turn dragging ability on/off, defaults False}

\item{.tap}{turn tap ability on/off, default False}
}
\description{
Use the standard JA base map layer in leaflet.
It includes OSM address search, easyPrint, and map label tiles on top.
You can also set the initial coordinates and zoom level for your map. Defaults to Harris County, TX.
One catch is that you must specify any heatmap layers with mapPane argument; see example below.
}
\examples{
ja_base_map() \%>\%
leaflet::addPolygons(
data = clean_tx_counties,
options = leaflet::leafletOptions(pane = "polygons")
)
}
