% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_geo_packages.R
\name{load_geo_packages}
\alias{load_geo_packages}
\title{Load geospatial packages}
\usage{
load_geo_packages()
}
\description{
Automatically loads geospatial suite of packages for analysis, including
sf, tigris, tidycensus, leaflet, and more.
}
\examples{
load_geo_packages()
}
