% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_geo_options.R
\name{set_geo_options}
\alias{set_geo_options}
\title{Set options of spatial analysis}
\usage{
set_geo_options()
}
\description{
Set the options for tigris and other spatial packages at start of analysis
}
\examples{
set_geo_options()
}
